snap: move AllUsers to snap
