+++ overlord/snapshotstate/backend/helpers.go
@@ -27,9 +27,7 @@ import (
 	"os/exec"
 	"os/user"
 	"path/filepath"
-	"strconv"
 	"strings"
-	"syscall"
 
 	"github.com/snapcore/snapd/client"
 	"github.com/snapcore/snapd/dirs"
@@ -97,7 +95,7 @@ var (
 
 func usersForUsernamesImpl(usernames []string, opts *dirs.SnapDirOptions) ([]*user.User, error) {
 	if len(usernames) == 0 {
-		return AllUsers(opts)
+		return snap.AllUsers(opts)
 	}
 	users := make([]*user.User, 0, len(usernames))
 	for _, username := range usernames {
@@ -130,59 +128,6 @@ func usersForUsernamesImpl(usernames []string, opts *dirs.SnapDirOptions) ([]*us
 		users = append(users, usr)
 
 	}
-	return users, nil
-}
-
-func AllUsers(opts *dirs.SnapDirOptions) ([]*user.User, error) {
-	ds, err := filepath.Glob(snap.DataHomeGlob(opts))
-	if err != nil {
-		// can't happen?
-		return nil, err
-	}
-
-	users := make([]*user.User, 1, len(ds)+1)
-	root, err := user.LookupId("0")
-	if err != nil {
-		return nil, err
-	}
-	users[0] = root
-	seen := make(map[uint32]bool, len(ds)+1)
-	seen[0] = true
-	var st syscall.Stat_t
-	for _, d := range ds {
-		err := syscall.Stat(d, &st)
-		if err != nil {
-			continue
-		}
-		if seen[st.Uid] {
-			continue
-		}
-		seen[st.Uid] = true
-		usr, err := userLookupId(strconv.FormatUint(uint64(st.Uid), 10))
-		if err != nil {
-			// Treat all non-nil errors as user.Unknown{User,Group}Error's, as
-			// currently Go's handling of returned errno from get{pw,gr}nam_r
-			// in the cgo implementation of user.Lookup is lacking, and thus
-			// user.Unknown{User,Group}Error is returned only when errno is 0
-			// and the list of users/groups is empty, but as per the man page
-			// for get{pw,gr}nam_r, there are many other errno's that typical
-			// systems could return to indicate that the user/group wasn't
-			// found, however unfortunately the POSIX standard does not actually
-			// dictate what errno should be used to indicate "user/group not
-			// found", and so even if Go is more robust, it may not ever be
-			// fully robust. See from the man page:
-			//
-			// > It [POSIX.1-2001] does not call "not found" an error, hence
-			// > does not specify what value errno might have in this situation.
-			// > But that makes it impossible to recognize errors.
-			//
-			// See upstream Go issue: https://github.com/golang/go/issues/40334
-			continue
-		} else {
-			users = append(users, usr)
-		}
-	}
-
 	return users, nil
 }
